Update to be a separate class instead of modifying core
